backend/requirements.txt
Flask
flask-cors
pymongo
bcrypt
Flask-Mail
python-dotenv
